backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
langgraph==0.0.20
langchain==0.1.0
chromadb>=0.4.22
sentence-transformers==2.2.2
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
typing-extensions==4.8.0
python-dotenv==1.0.0
